Convert decision tree building for new deductions
The build-time data structure now support datapoints created later
in the building process. These datapoints get labeled or eventually
processed in their node.

While we build the decision tree, we keep track of whether each
datapoint has been split (ended up on both sides of a decision) or
is whole. If the datapoint is split, only the constraints related to
the whole datapoint matter in labeling, because all of its subsets
which are relevant to some constraint have already been created.